+++ src/core/basetypes/MCString.cc
@@ -2125,8 +2125,8 @@ bool String::isEqualCaseInsensitive(String * otherString)
 Data * String::decodedBase64Data()
 {
     const char * utf8 = UTF8Characters();
-    char * decoded = MCDecodeBase64(utf8, strlen(utf8));
-    Data * result = Data::dataWithBytes(decoded, strlen(decoded));
+    char * decoded = MCDecodeBase64(utf8, (unsigned int) strlen(utf8));
+    Data * result = Data::dataWithBytes(decoded, (unsigned int) strlen(decoded));
     free(decoded);
     return result;
 }
